Day 4: Java workshop progress

+++ week-01/day-3/src/oddEven.java
@@ -5,7 +5,7 @@ public class oddEven {
         // Write a program that reads a number from the standard input,
         // Then prints "Odd" if the number is odd, or "Even" if it is even.
         Scanner scanner = new Scanner(System.in);
-        int userInputNumber = (int) scanner.nextInt();
+        long userInputNumber = (long) scanner.nextInt();
 
         if (userInputNumber % 2 == 0) {
             System.out.println("Even");
